Skip packages that have already been expired.
If a package is removed, then expire-pool.py will remove our copies of
its source code; if it is later reintroduced, then update-pool.py won't
download it again because we've already processed it, but
generate-diffs.py and other scripts will crash because they see it in
indexes but can't find the source code.  Just skip packages in this
state, since we probably already have suitable diffs for them.