remove_old_comments: always seek(0) before reading the comments file
The symptom of not doing so was the comment file emptying at each run of
`manual-status.py`.

Here is what `man 3 fopen` says:
```
 a+     Open for reading and appending (writing at end of file).  The file is created  if  it  does  not  exist.
        Output  is always appended to the end of the file.  POSIX is silent on what the initial read position is
        when using this mode.  For glibc, the initial file position for reading is at the beginning of the file,
        but for Android/BSD/MacOS, the initial file position for reading is at the end of the file.
```

Since this is not specified, obviously, there is a change of behavior in the Python history:
```bash
root@jammy:/tmp# python2.7 test.py
with open("test.py", "a+") as f:

    for line in f:

        print(line)

root@jammy:/tmp# python3.10 test.py
root@jammy:/tmp#
```
At least the following bug from Python 3.4 era exists on the topic:
https://bugs.python.org/issue22651

Digging exactly when this change happened is left as an exercise to
the reader.