momlib: set ROOT to reflect what's need for production
This is needed because the code handling the comments is run by apache,
meaning with the www-data user.
This is less than ideal, but it's also all we have right now, and doing
better requires more work. Ideally, we want to square out all the path
handling to have something more contained when starting from a `git
clone`.